requirements.txt
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz
pynput
pystray